Remove previous notification before showing new one
Windows does not show the new notification otherwise. I learnt this the
hard way - if two notifications appear successively, the second one was
not being displayed. If the second notification is sent after the first
notification is closed, then it gets displayed. Fix is to remove the old
notification and show new one